change variable type of id

+++ src/main/java/weather/server/controller/CityController.java
@@ -3,5 +3,5 @@ package weather.server.controller;
 import weather.server.to.CityTO;
 
 public interface CityController {
-    public CityTO findCityById(Integer id);
+    public CityTO findCityById(Long id);
 }
